🏁 Milestone: the entire Go standard library compiles — 302/302, zero errors
All 302 packages of the auto-converted Go standard library (Go 1.23.1, see
go-src-converted/VERSION) compile cleanly as .NET assemblies. This commit is
the milestone artifact:

- docs/README.md gains the NEWS section telling the story — eight years of
  human design, architecture, and foundation; two weeks and 939 commits of
  AI campaign (683 individually-gated converter/runtime/generator fixes,
  each locked by one of 371 Go-vs-C# behavioral regression tests) — and
  fills the Milestones row this table has been holding open.
- src/go-src-converted is the verified compiled snapshot: the exact output
  of the milestone converter (master 51ba5d9cf), overlaid per the Phase-3
  mechanics (fresh csprojs with go-src-converted references, hand-owned
  manual implementations restored), census-proven 302/302 with zero own
  errors anywhere in the corpus. Committed deliberately as the milestone
  exception to the regens-are-not-committed rule.
- go-src-converted/VERSION records the Go toolchain (hand-copied this once;
  src/go2cs/ToDo.md notes automating it into the -stdlib output).

Compiling is the milestone; operational is the mission. Phase 4 — running
Go's own test suites against this output — starts from here.

+++ src/go-src-converted/compress/flate/token.cs
@@ -65,41 +65,41 @@ internal static array<uint32> offsetCodes = new uint32[]{
 
 // Convert a literal into a literal token.
 internal static token literalToken(uint32 literal) {
-    return ((token)(literalType + literal));
+    return ((token)((uint32)literalType + literal));
 }
 
 // Convert a < xlength, xoffset > pair into a match token.
 internal static token matchToken(uint32 xlength, uint32 xoffset) {
-    return ((token)(matchType + xlength << (int)(lengthShift) + xoffset));
+    return ((token)((uint32)matchType + (xlength << (int)(lengthShift)) + xoffset));
 }
 
 // Returns the literal of a literal token.
 internal static uint32 literal(this token t) {
-    return ((uint32)(t - literalType));
+    return (uint32)(t - (uint32)literalType);
 }
 
 // Returns the extra offset of a match token.
 internal static uint32 offset(this token t) {
-    return (uint32)(((uint32)t) & offsetMask);
+    return (uint32)((uint32)t & (uint32)offsetMask);
 }
 
 internal static uint32 length(this token t) {
-    return ((uint32)((t - matchType) >> (int)(lengthShift)));
+    return (uint32)(((t - (uint32)matchType) >> (int)(lengthShift)));
 }
 
 internal static uint32 lengthCode(uint32 len) {
-    return lengthCodes[len];
+    return lengthCodes[(nint)(len)];
 }
 
 // Returns the offset code corresponding to a specific offset.
 internal static uint32 offsetCode(uint32 off) {
-    if (off < ((uint32)len(offsetCodes))) {
-        return offsetCodes[off];
+    if (off < (uint32)len(offsetCodes)) {
+        return offsetCodes[(nint)(off)];
     }
-    if (off >> (int)(7) < ((uint32)len(offsetCodes))) {
-        return offsetCodes[off >> (int)(7)] + 14;
+    if ((off >> (int)(7)) < (uint32)len(offsetCodes)) {
+        return offsetCodes[(nint)((off >> (int)(7)))] + 14;
     }
-    return offsetCodes[off >> (int)(14)] + 28;
+    return offsetCodes[(nint)((off >> (int)(14)))] + 28;
 }
 
 } // end flate_package
