+++ src/go-src-converted/internal/itoa/itoa.cs
@@ -9,9 +9,9 @@ partial class itoa_package {
 // Itoa converts val to a decimal string.
 public static @string Itoa(nint val) {
     if (val < 0) {
-        return "-"u8 + Uitoa(((nuint)(-val)));
+        return "-"u8 + Uitoa((nuint)(-val));
     }
-    return Uitoa(((nuint)val));
+    return Uitoa((nuint)val);
 }
 
 // Uitoa converts val to a decimal string.
@@ -24,12 +24,12 @@ public static @string Uitoa(nuint val) {
     nint i = len(buf) - 1;
     while (val >= 10) {
         nuint q = val / 10;
-        buf[i] = ((byte)((rune)'0' + val - q * 10));
+        buf[i] = (byte)((rune)'0' + val - q * 10);
         i--;
         val = q;
     }
     // val < 10
-    buf[i] = ((byte)((rune)'0' + val));
+    buf[i] = (byte)((rune)'0' + val);
     return ((@string)(buf[(int)(i)..]));
 }
 
@@ -45,12 +45,12 @@ public static @string Uitox(nuint val) {
     nint i = len(buf) - 1;
     while (val >= 16) {
         nuint q = val / 16;
-        buf[i] = hex[val % 16];
+        buf[i] = hex[(int)(val % 16)];
         i--;
         val = q;
     }
     // val < 16
-    buf[i] = hex[val % 16];
+    buf[i] = hex[(int)(val % 16)];
     i--;
     buf[i] = (rune)'x';
     i--;
