+++ src/go-src-converted/net/sock_windows.cs
@@ -18,8 +18,8 @@ internal static nint maxListenerBacklog() {
 }
 
 internal static (syscallꓸHandle, error) sysSocket(nint family, nint sotype, nint proto) {
-    var (s, err) = wsaSocketFunc(((int32)family), ((int32)sotype), ((int32)proto),
-        nil, 0, (uint32)(windows.WSA_FLAG_OVERLAPPED | windows.WSA_FLAG_NO_HANDLE_INHERIT));
+    var (s, err) = wsaSocketFunc((int32)family, (int32)sotype, (int32)proto,
+        nil, 0, (uint32)((uint32)windows.WSA_FLAG_OVERLAPPED | (uint32)windows.WSA_FLAG_NO_HANDLE_INHERIT));
     if (err != default!) {
         return (syscall.InvalidHandle, os.NewSyscallError("socket"u8, err));
     }
